Lazy load the FieldArray as it takes at least 3 seconds to load
