+++ src/components/Layouts/Primary/TopBar/Items/AddMenu/Items/CreateMultipleContacts/DynamicCreateMultipleContactsFieldArray.tsx
@@ -1,0 +1,10 @@
+import dynamic from 'next/dynamic';
+import { DynamicComponentPlaceholder } from 'src/components/DynamicPlaceholders/DynamicComponentPlaceholder';
+
+export const DynamicContactsRightPanel = dynamic(
+  () =>
+    import(
+      /* webpackChunkName: "ContactsRightPanel" */ './CreateMultipleContactsFieldArray'
+    ),
+  { loading: DynamicComponentPlaceholder },
+);
